server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openpyxl>=3.1.0
python-multipart>=0.0.6
# Auth: sign-in / JWT
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# Real AI: CLIP zero-shot image classification
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
